requirements.txt
flask==3.1.1             # web API
python-dotenv==1.1.1     # load env
tensorflow==2.19.0       # ML
prometheus_client==0.22.1 # metrics
numpy==1.26.4             # data prep
pytest==8.4.0            # testing
joblib==1.5.1            # model serialization

psycopg2-binary==2.9.9       # postgres client
